pyproject.toml
[project]
name = "open-ghana-id"
version = "0.1.0"
description = "Open-source pre-verification service for Ghanaian IDs (educational use only)."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Jerry Buaba" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "opencv-python-headless>=4.10.0.0",
  "pillow>=11.0.0",
  "numpy>=2.1.0",
  "passporteye<=2.2.2",
  "pytesseract>=0.3.13",
  "aiofiles>=24.1.0",
  "qrcode>=7.4.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "types-requests>=2.32.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/open-ghana-id"

[tool.uv]
dev-dependencies = [
  "pytest",
  "pytest-asyncio",
  "ruff",
  "mypy",
  "types-requests",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
testpaths = ["tests"]

